Fix some POSIX issues in install.sh

+++ install.sh
@@ -18,13 +18,13 @@ fi
 
 usage() {
     USAGE="Usage: $0"
-    SPACE=$(echo ${USAGE} | tr "[:print:]" " ")
-    echo ${C_INFO}
+    SPACE=$(echo "${USAGE}" | tr "[:print:]" " ")
+    echo "${C_INFO}"
     echo "${USAGE} [-i] [-u] [-d] [-z] [-s] [-t]"
     echo "${SPACE} [-b <branch>]"
     echo "${SPACE} [-k <klipper_home_dir>] [-c <klipper_config_dir>] [-m <moonraker_home_dir>]"
     echo "${SPACE} [-a <kiauh_alternate_klipper>] [config_file]" # [-r <repetier_server stub>]"
-    echo ${C_OFF}
+    echo "${C_OFF}"
     echo "${C_INFO}(no flags for safe re-install / upgrade)${C_OFF}"
     echo "-i for interactive install"
     echo "-u or -d for uninstall"
@@ -40,7 +40,7 @@ usage() {
     echo "-a <name> to specify alternative klipper-service-name when installed with Kiauh"
     echo "-t activate test mode to create test config files in /tmp"
     echo "${C_INFO}[config_file]${C_OFF} is optional, if not specified the default config filename (.config) will be used."
-    echo "(-q verbose make; -v verbose builder)"
+    echo "(-q verbose make --no-print-directory; -v verbose builder)"
     echo
     exit 0
 }
@@ -56,15 +56,14 @@ ordinal() {
 
 prompt_yn() {
     while true; do
-        read -n1 -p "$* (y/n)? " yn
+        printf "%s (y/n)? " "$*"
+        read -r yn
         case "${yn}" in
         Y | y)
-            echo -n "y"
-            break
+            return 1
             ;;
         N | n)
-            echo -n "n"
-            break
+            return 0
             ;;
         esac
     done
@@ -105,7 +104,7 @@ done
 # Handle git self update or branch change
 if [ "${F_SKIP_UPDATE}" = "force" ]; then
     : # If we just restarted with a forced skip, do nothing
-elif [ ! "${F_SKIP_UPDATE}" ]; then
+elif [ ! "${F_SKIP_UPDATE}" ] && [ ! "${F_UNINSTALL}" ]; then
     [ -t 1 ] && clear
     "$SCRIPT_DIR/installer/self_update.sh" || exit 1
     F_SKIP_UPDATE=force exec "$0" "$@"
@@ -132,19 +131,22 @@ if [ "${TESTDIR}" ]; then
     export CONFIG_MOONRAKER_HOME="${TESTDIR}/moonraker"
     export F_NO_SERVICE=y
     export KCONFIG_CONFIG="${TESTDIR}/.config"
-    echo -e "\n${C_WARNING}Running in test mode to simulate without changing real configuration${C_OFF}"
-    echo -e "${C_WARNING}Forcing flags '-s -c ${CONFIG_KLIPPER_CONFIG_HOME} -k ${CONFIG_KLIPPER_HOME} -c ${CONFIG_MOONRAKER_HOME} ${TESTDIR}/.config' ${C_OFF}"
+    echo
+    echo "${C_WARNING}Running in test mode to simulate without changing real configuration${C_OFF}"
+    echo "${C_WARNING}Forcing flags '-s -c ${CONFIG_KLIPPER_CONFIG_HOME} -k ${CONFIG_KLIPPER_HOME} -c ${CONFIG_MOONRAKER_HOME} ${TESTDIR}/.config' ${C_OFF}"
     mkdir -p "${CONFIG_KLIPPER_HOME}/klippy/extras"
     mkdir -p "${CONFIG_KLIPPER_CONFIG_HOME}"
     mkdir -p "${CONFIG_MOONRAKER_HOME}/moonraker/components"
     touch "${CONFIG_KLIPPER_CONFIG_HOME}/moonraker.conf"
     touch "${CONFIG_KLIPPER_CONFIG_HOME}/printer.cfg"
     if [ ! "${F_UNINSTALL}" ]; then
-        echo -e "${C_INFO}When complete look in ${TESTDIR} for results${C_OFF}\n"
-        [ "$(prompt_yn "Continue")" != "y" ] && {
+        echo "${C_INFO}When complete look in ${TESTDIR} for results${C_OFF}"
+        echo
+        if prompt_yn "Continue"; then
             echo
             exit 0
-        } || echo
+        fi
+        echo
         echo
     fi
 fi
@@ -154,16 +156,16 @@ fi
 #####################
 
 if [ "${F_UNINSTALL}" ]; then
-    echo -e "${C_WARNING}This will uninstall and cleanup prior config${C_OFF}\n"
-    [ "$(prompt_yn "Are you sure")" != "y" ] && {
-        echo
-        exit 0
-    } || echo
+    echo "${C_WARNING}This will uninstall and cleanup prior config${C_OFF}"
     echo
-    SECONDS=0
-    make -C "${SCRIPT_DIR}" uninstall
-    [ "${TESTDIR}" ] && rm -rf "${TESTDIR}"
-    echo "${C_INFO}Elapsed: ${SECONDS} seconds${C_OFF}"
+    if prompt_yn "Are you sure"; then
+        exit 0
+    fi
+    echo
+    echo
+    command time -f "${C_INFO}Elapsed: %e seconds${C_OFF}" \
+        make --no-print-directory -C "${SCRIPT_DIR}" uninstall &&
+        [ "${TESTDIR}" ] && rm -rf "${TESTDIR}"
     exit 0
 fi
 
@@ -190,9 +192,9 @@ if [ -n "${F_MENUCONFIG:-}" ]; then
             tmpconfig="$(mktemp -t tmpconfig.XXXXXX)"
             cp -- "${KCONFIG_CONFIG}" "${tmpconfig}"
         fi
-        make -C "${SCRIPT_DIR}" F_MULTI_UNIT_ENTRY_POINT=y F_MULTI_UNIT=y menuconfig
+        make --no-print-directory -C "${SCRIPT_DIR}" F_MULTI_UNIT_ENTRY_POINT=y F_MULTI_UNIT=y menuconfig
     else
-        make -C "${SCRIPT_DIR}" menuconfig
+        make --no-print-directory -C "${SCRIPT_DIR}" menuconfig
     fi
 
     if [ ! -e "${KCONFIG_CONFIG}" ]; then
@@ -218,7 +220,7 @@ if [ -n "${F_MENUCONFIG:-}" ]; then
         for name in ${CONFIG_PARAM_MMU_UNITS:-}; do
             name=$(trim "$name")
             [ -n "$name" ] || continue
-            make -C "${SCRIPT_DIR}" KCONFIG_CONFIG="${KCONFIG_CONFIG}_${name}" F_MULTI_UNIT=y UNIT_INDEX="${i}" UNIT_NAME="${name}" MMU_MCU="${name}" menuconfig
+            make --no-print-directory -C "${SCRIPT_DIR}" KCONFIG_CONFIG="${KCONFIG_CONFIG}_${name}" F_MULTI_UNIT=y UNIT_INDEX="${i}" UNIT_NAME="${name}" MMU_MCU="${name}" menuconfig
             i=$((i + 1))
         done
         set +f
@@ -230,6 +232,5 @@ fi
 ##### Install #####
 ###################
 
-SECONDS=0
-make -C "${SCRIPT_DIR}" KCONFIG_CONFIG="${KCONFIG_CONFIG}" install
-echo "${C_INFO}Elapsed: ${SECONDS} seconds${C_OFF}"
+command time -f "${C_INFO}Elapsed: %e seconds${C_OFF}" \
+    make --no-print-directory -C "${SCRIPT_DIR}" KCONFIG_CONFIG="${KCONFIG_CONFIG}" install
